Full Authorization

+++ App/controllers/ListingController.php
@@ -3,7 +3,9 @@
 namespace App\Controllers;
 
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -18,7 +20,7 @@ class ListingController
 
     public function index()
     {
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY create_at DESC')->fetchAll();
 
         loadView('listings/index', ['listings' => $listings]);
     }
@@ -36,7 +38,7 @@ class ListingController
             'id' => $id
         ];
 
-        $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
+        $listing = $this->db->query('SELECT * FROM  listings WHERE id = :id', $params)->fetch();
 
         // Check if listing exsist
         if (!$listing) {
@@ -53,7 +55,7 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -123,10 +125,15 @@ class ListingController
             return;
         }
 
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessages('error_message', 'You are not authorized to delete');
+            return redirect('/listings/' . $listing->id);
+        }
+
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
         // Set flash message
-        $_SESSION['success_message'] = 'Listing delete successfully';
+        Session::setFlashMessages('success_message', 'Listing delete successfully');
 
         redirect('/listings');
     }
@@ -145,6 +152,12 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessages('error_message', 'You are not authoirzed to update this listing');
+            return redirect('/listings/' . $listing->id);
         }
 
         loadView('listings/edit', ['listing' => $listing]);
@@ -167,10 +180,10 @@ class ListingController
         }
 
         // Authorization
-        // if (!Authorization::isOwner($listing->user_id)) {
-        //     Session::setFlashMessage('error_message', 'You are not authoirzed to update this listing');
-        //     return redirect('/listings/' . $listing->id);
-        // }
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessages('error_message', 'You are not authoirzed to update this listing');
+            return redirect('/listings/' . $listing->id);
+        }
 
         $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits'];
 
@@ -212,7 +225,7 @@ class ListingController
             $this->db->query($updateQuery, $updateValues);
 
             // Set flash message
-            // Session::setFlashMessage('success_message', 'Listing updated');
+            Session::setFlashMessages('success_message', 'Listing updated');
 
             redirect('/listings/' . $id);
         }
